requirements.txt
requests
python-dotenv
pandas
numpy
scikit-learn
tensorflow
mlflow
fastapi
uvicorn
streamlit
python-multipart
prometheus-client
